move the real transport components into the 'real' submodule
