+++ dnp3/src/transport/real/assembler.rs
@@ -1,5 +1,5 @@
 use crate::link::header::FrameInfo;
-use crate::transport::header::Header;
+use crate::transport::real::header::Header;
 use crate::transport::{Fragment, FragmentInfo};
 use crate::util::buffer::Buffer;
