+++ dnp3/src/transport/real/mod.rs
@@ -1,0 +1,5 @@
+pub(crate) mod assembler;
+pub(crate) mod header;
+pub(crate) mod reader;
+pub(crate) mod sequence;
+pub(crate) mod writer;
